+++ piping-webrtc-signaling/answer.go
@@ -6,83 +6,123 @@ import (
 	"github.com/pion/webrtc/v3"
 	"log"
 	"net/http"
+	"net/url"
 	"sync"
 	"time"
 )
 
 type Answer struct {
-	pipingServerUrl   string
-	httpHeaders       [][]string
-	peerConnection    *webrtc.PeerConnection
-	answerSideId      string
-	offerSideId       string
-	logger            *log.Logger
-	httpClient        *http.Client
-	candidatesMux     sync.Mutex
-	pendingCandidates []*webrtc.ICECandidate
+	pipingServerUrl *url.URL
+	httpHeaders     [][]string
+	peerConnection  *webrtc.PeerConnection
+	answerSideId    string
+	offerSideId     string
+	logger          *log.Logger
+	httpClient      *http.Client
 }
 
-func NewAnswer(logger *log.Logger, httpClient *http.Client, pipingServerUrl string, httpHeaders [][]string, peerConnection *webrtc.PeerConnection, answerSideId string, offerSideId string) *Answer {
-	return &Answer{
-		pipingServerUrl:   pipingServerUrl,
-		httpHeaders:       httpHeaders,
-		peerConnection:    peerConnection,
-		answerSideId:      answerSideId,
-		offerSideId:       offerSideId,
-		logger:            logger,
-		httpClient:        httpClient,
-		candidatesMux:     sync.Mutex{},
-		pendingCandidates: make([]*webrtc.ICECandidate, 0),
+func NewAnswer(logger *log.Logger, httpClient *http.Client, pipingServerUrlStr string, httpHeaders [][]string, peerConnection *webrtc.PeerConnection, answerSideId string, offerSideId string) (*Answer, error) {
+	pipingServerUrl, err := url.Parse(pipingServerUrlStr)
+	if err != nil {
+		return nil, err
 	}
+	return &Answer{
+		pipingServerUrl: pipingServerUrl,
+		httpHeaders:     httpHeaders,
+		peerConnection:  peerConnection,
+		answerSideId:    answerSideId,
+		offerSideId:     offerSideId,
+		logger:          logger,
+		httpClient:      httpClient,
+	}, nil
 }
 
 func (a *Answer) Start() error {
 	errCh := make(chan error)
 
+	candidatesMux := sync.Mutex{}
+	pendingCandidates := make([]*webrtc.ICECandidate, 0)
+	candidateFinished := false
+	notifiedCandidateFinish := false
+
 	a.peerConnection.OnICECandidate(func(c *webrtc.ICECandidate) {
+		candidatesMux.Lock()
+		defer candidatesMux.Unlock()
+
+		desc := a.peerConnection.RemoteDescription()
+
 		if c == nil {
+			candidateFinished = true
+			if desc == nil {
+				return
+			}
+			if err := a.sendCandidates([]*webrtc.ICECandidate{}); err != nil {
+				errCh <- err
+				return
+			}
+			notifiedCandidateFinish = true
 			return
 		}
 
-		a.candidatesMux.Lock()
-		defer a.candidatesMux.Unlock()
-
-		desc := a.peerConnection.RemoteDescription()
 		if desc == nil {
-			a.pendingCandidates = append(a.pendingCandidates, c)
-		} else if err := a.sendCandidate(c); err != nil {
+			pendingCandidates = append(pendingCandidates, c)
+		} else if err := a.sendCandidates([]*webrtc.ICECandidate{c}); err != nil {
 			errCh <- err
 		}
 	})
 
-	var initial InitialJson
+	var offerInitial OfferInitialJson
 	for {
-		res, err := a.httpClient.Get(fmt.Sprintf("%s/%s-%s", a.pipingServerUrl, a.offerSideId, a.answerSideId))
+		err := func() error {
+			offerInitialBytes, err := httpGetWithHeaders(a.httpClient, urlJoin(a.pipingServerUrl, sha256String(fmt.Sprintf("%s-%s", a.offerSideId, a.answerSideId))), a.httpHeaders)
+			if err != nil {
+				return err
+			}
+			if err = json.Unmarshal(offerInitialBytes, &offerInitial); err != nil {
+				return err
+			}
+			return nil
+		}()
 		if err != nil {
-			goto retry
-		}
-		if err = json.NewDecoder(res.Body).Decode(&initial); err != nil {
-			goto retry
+			a.logger.Printf("error: %+v", err)
+			time.Sleep(3 * time.Second)
+			continue
 		}
 		break
-	retry:
-		a.logger.Printf("error: %+v", err)
-		time.Sleep(3 * time.Second)
 	}
-	a.logger.Printf("initial: %+v", initial)
+	a.logger.Printf("offerInitial: %+v", offerInitial)
+
+	answerInitial := AnswerInitialJson{Version: 1}
+	answerInitialBytes, err := json.Marshal(answerInitial)
+	if err != nil {
+		return err
+	}
+	for {
+		err := pipingPostJson(a.httpClient, urlJoin(a.pipingServerUrl, sha256String(fmt.Sprintf("%s-%s", a.answerSideId, a.offerSideId))), a.httpHeaders, answerInitialBytes)
+		if err != nil {
+			a.logger.Printf("failed to send answerInitial: %+v", err)
+			time.Sleep(3 * time.Second)
+			continue
+		}
+		break
+	}
 
 	go func() {
-		// TODO: finish loop when connected
 		for {
-			candidate, err := receiveCandidate(a.httpClient, a.pipingServerUrl, a.httpHeaders, a.answerSideId, a.offerSideId)
+			candidates, err := receiveCandidates(a.httpClient, a.pipingServerUrl, a.httpHeaders, a.answerSideId, a.offerSideId)
 			if err != nil {
 				errCh <- err
 				return
 			}
+			if len(candidates) == 0 {
+				break
+			}
 			a.logger.Printf("candidate received")
-			if err := a.peerConnection.AddICECandidate(*candidate); err != nil {
-				errCh <- err
-				return
+			for _, candidate := range candidates {
+				if err := a.peerConnection.AddICECandidate(candidate); err != nil {
+					errCh <- err
+					return
+				}
 			}
 		}
 	}()
@@ -124,23 +164,29 @@ func (a *Answer) Start() error {
 			errCh <- err
 			return
 		}
-		a.candidatesMux.Lock()
-		defer a.candidatesMux.Unlock()
-		for _, c := range a.pendingCandidates {
+		candidatesMux.Lock()
+		defer candidatesMux.Unlock()
+		if len(pendingCandidates) != 0 {
 			for {
-				if err = a.sendCandidate(c); err != nil {
-					a.logger.Printf("failed to send candidate: %+v", err)
+				if err = a.sendCandidates(pendingCandidates); err != nil {
+					a.logger.Printf("failed to send candidates: %+v", err)
 					time.Sleep(3 * time.Second)
 					continue
 				}
 				break
 			}
 		}
+		if candidateFinished && !notifiedCandidateFinish {
+			if err := a.sendCandidates([]*webrtc.ICECandidate{}); err != nil {
+				errCh <- err
+			}
+			notifiedCandidateFinish = true
+		}
 	}()
 
 	return <-errCh
 }
 
-func (a *Answer) sendCandidate(candidate *webrtc.ICECandidate) error {
-	return sendCandidate(a.logger, a.httpClient, a.pipingServerUrl, a.httpHeaders, a.answerSideId, a.offerSideId, candidate)
+func (a *Answer) sendCandidates(candidates []*webrtc.ICECandidate) error {
+	return sendCandidates(a.logger, a.httpClient, a.pipingServerUrl, a.httpHeaders, a.answerSideId, a.offerSideId, candidates)
 }