+++ tunnel/dialer.go
@@ -1,10 +1,10 @@
 // base: https://github.com/pion/webrtc/tree/80e5cdda5687d696556f2f2605a4c83f61ac3a08/examples/pion-to-pion
 
-package core
+package tunnel
 
 import (
 	"fmt"
-	piping_webrtc_signaling "github.com/nwtgck/go-webrtc-piping-tunnel/piping-webrtc-signaling"
+	piping_webrtc_signaling "github.com/nwtgck/go-webrtc-piping/piping-webrtc-signaling"
 	"github.com/pion/webrtc/v3"
 	"io"
 	"log"
@@ -58,7 +58,11 @@ func Dialer(logger *log.Logger, httpClient *http.Client, pipingServerUrl string,
 	}
 
 	go func() {
-		answer := piping_webrtc_signaling.NewAnswer(logger, httpClient, pipingServerUrl, httpHeaders, peerConnection, answerSideId(path), offerSideId(path))
+		answer, err := piping_webrtc_signaling.NewAnswer(logger, httpClient, pipingServerUrl, httpHeaders, peerConnection, answerSideId(path), offerSideId(path))
+		if err != nil {
+			errCh <- err
+			return
+		}
 		if err := answer.Start(); err != nil {
 			errCh <- err
 		}