+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.3.0] - 2023-03-19
+### Changed
+* Rename webrtc-piping-tunnel → webrtc-piping
+* Create "tunnel" and "duplex" subcommands
+* Update dependencies
+
 ## [0.2.0] - 2022-07-04
 ### Added
 * Support UDP port forwarding
@@ -16,5 +22,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Added
 * Initial release
 
-[Unreleased]: https://github.com/nwtgck/go-webrtc-piping-tunnel/compare/v0.2.0...HEAD
-[0.2.0]: https://github.com/nwtgck/go-webrtc-piping-tunnel/compare/v0.1.0...v0.2.0
+[Unreleased]: https://github.com/nwtgck/go-webrtc-piping/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/nwtgck/go-webrtc-piping/compare/v0.2.0...v0.3.0
+[0.2.0]: https://github.com/nwtgck/go-webrtc-piping/compare/v0.1.0...v0.2.0